[Distributed] Do not use nlohmann/json.h in public headers
Move dependency introduced by https://github.com/pytorch/pytorch/pull/129505 into .cpp file

[ghstack-poisoned]
